// src/components/createTaskForm/interfaces/ISelectField.ts
import { IDisabled } from './IDisabled';
import { SelectChangeEvent } from '@mui/material';

export interface iSelectItems {
  value: string;
  label: string;
}

export interface ISelectField extends IDisabled {
  name?: string;
  label?: string;
  value?: string;
  onChange?: (e: SelectChangeEvent) => void;
  items?: iSelectItems[];
}